vibe coded some unit tests, which i should have been writing already, also lighting wasnt being applied properly when scaling spheres, becouse the position function was taking int time instead of float which was too inaccurate and probably always rounded to 1/0 when working with smaller spheres

+++ includes/rays.h
@@ -6,7 +6,7 @@
 /*   By: bpetrovi <bpetrovi@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/30 18:56:37 by bpetrovi          #+#    #+#             */
-/*   Updated: 2026/06/04 21:13:01 by bpetrovi         ###   ########.fr       */
+/*   Updated: 2026/06/08 14:55:39 by bpetrovi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ typedef struct s_sphere
 t_ray		ray(t_tuple origin, t_tuple direction);
 t_ray		ray_transform(t_ray ray, t_matrix transformation);
 t_sphere	sphere(int id);
-t_tuple		position(t_ray ray, int time);
+t_tuple		position(t_ray ray, float time);
 void		set_transformation(t_sphere *sphere, t_matrix transformation);
